Rename files in string-escape package
