+++ packages/@romejs/string-escape/escapeJSString.test.ts
